refactor: extract reusable functions from CustomerCalendar to calendarUtils.tsx
